Don't set CORS headers in nginx server
ga4gh/dockstore#2056

The Java server already sets the the CORS headers using
CrossOriginFilter in DockstoreWebserviceApplication.java. Setting the
access-control-allow-origin-header to two values is not valid --
that was happening because the headers were being set twice.

The reason that this bug only occurred on port 443 and not on port
8443 is because the CORS headers were only set for the server on
port 4200, which maps to port 443 in docker-compose.yml.

The server section for port 8080, which maps to port 8443, was
and isn't setting the CORS headers.

Because only AJAX requests bring CORS into the equation for Dockstore,
see https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#What_requests_use_CORS,
this should be safe to remove from nginx, as the only thing that
implements an API that AJAX requests can invoke, is the
webservice, which already handles CORS.