Handle ampersand values and add 1.11 migration
* Handle password with ampersand

dockstore/dockstore#3971

There is a password with an ampersand in an environment.
It caused two problems:

1. Ampersand was getting escaped in init_migration.sh.template;
Use triple braces to avoid
2. The jq logic in install_bootstrap converts the JSON file
to a bunch of bash variable assignments executed with
source. Because a value with
an ampersand was not quoted, it was splitting the value into
two commands, which failed. Double quote the values.

Noticed the missing migration to 1.11 when working on this.